added cookie-based IRestAuthClient and moved IRest____Client items to Sienar.Infrastructure namespace
