+++ src/Sienar.Architecture.Rest/Infrastructure/IRestClient.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Sienar.Data;
 
-namespace Sienar.Services;
+namespace Sienar.Infrastructure;
 
 public interface IRestClient
 {
